Refactor translation helper code and implement translation vcd test.
