+++ experiments/translation/exp_translation.py
@@ -1,237 +1,4 @@
-import numpy as np
-from PIL import Image, ImageDraw, ImageFont
-
 import mbirjax as mj
-import matplotlib.pyplot as plt
-
-
-# Utility function to display translation vectors' x and z components using a scatter plot
-def display_translation_vectors(translation_vectors, recon_shape):
-    """Display the x and z components of translation vectors using a scatter plot,
-    and overlay a box representing the reconstruction volume in (column, slice) space.
-
-    Args:
-        translation_vectors (np.ndarray): Array of shape (N, 3) containing [dx, dy, dz] vectors.
-        recon_shape (tuple[int, int, int]): Shape of the reconstruction volume (rows, columns, slices).
-    """
-    dx = translation_vectors[:, 0]
-    dz = translation_vectors[:, 2]
-
-    plt.figure(figsize=(6, 6))
-    plt.scatter(dx, dz, c='blue', marker='o', label='Translations')
-
-    # Get col and slice dimensions (horizontal and vertical axes in view)
-    num_cols = recon_shape[1]
-    num_slices = recon_shape[2]
-
-    # Compute box boundaries centered around origin
-    half_width = num_cols / 2
-    half_height = num_slices / 2
-
-    box_x = [-half_width, half_width, half_width, -half_width, -half_width]
-    box_z = [-half_height, -half_height, half_height, half_height, -half_height]
-
-    plt.plot(box_x, box_z, 'r--', linewidth=2, label='Reconstruction Region')
-
-    plt.title("Translation Grid Points with Recon Outline")
-    plt.xlabel("Horizontal Translation in ALU")
-    plt.ylabel("Vertical Translation in ALU")
-    plt.axis('equal')
-    plt.legend(loc='upper right')
-    plt.show()
-
-
-def gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing):
-    """
-    Generate translation vectors for lateral (x) and axial (z) displacements.
-
-    Args:
-        num_x_translations (int): Number of x-direction translations
-        num_z_translations (int): Number of z-direction translations
-        x_spacing (float): Spacing between x translations in ALU
-        z_spacing (float): Spacing between z translations in ALU
-
-    Returns:
-        np.ndarray: Array of shape (num_views, 3) with translation vectors [dx, dy, dz]
-    """
-    num_views = num_x_translations * num_z_translations
-    translation_vectors = np.zeros((num_views, 3))
-
-    x_center = (num_x_translations - 1) / 2
-    z_center = (num_z_translations - 1) / 2
-
-    idx = 0
-    for row in range(num_z_translations):
-        for col in range(num_x_translations):
-            dx = (col - x_center) * x_spacing
-            dz = (row - z_center) * z_spacing
-            dy = 0
-            translation_vectors[idx] = [dx, dy, dz]
-            idx += 1
-
-    return translation_vectors
-
-
-def save_volume_as_gif(volume, filename, vmin=0, vmax=1):
-    """
-    Save a 3D volume as a GIF, iterating over axis 0 (row-wise).
-
-    Args:
-        volume (np.ndarray): 3D array to save as a movie.
-        filename (str): Output path for the GIF file.
-        vmin (float): Min pixel value for display normalization.
-        vmax (float): Max pixel value for display normalization.
-    """
-    try:
-        import imageio.v2 as imageio
-    except ImportError:
-        print("The 'imageio' package is not installed. Please install it using:\n    pip install imageio")
-        return
-
-    from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
-    images = []
-    for i in range(volume.shape[0]):
-        fig, ax = plt.subplots()
-        canvas = FigureCanvas(fig)
-        ax.imshow(volume[i, :, :].T, cmap='gray', vmin=vmin, vmax=vmax)
-        ax.axis('off')
-        # Convert canvas to image using RGBA buffer, then drop alpha channel
-        canvas.draw()
-        buf = canvas.get_renderer().buffer_rgba()
-        image = np.frombuffer(buf, dtype=np.uint8).reshape(canvas.get_width_height()[::-1] + (4,))
-        image = image[..., :3]  # Drop alpha channel
-        images.append(image)
-        plt.close(fig)
-
-    imageio.mimsave(filename, images, fps=5)  # 5 frames per second
-
-def gen_dot_phantom(recon_shape):
-    """
-    Generate a synthetic ground truth reconstruction volume.
-
-    Args:
-        recon_shape (tuple[int, int, int]): Shape of the reconstruction volume.
-
-    Returns:
-        np.ndarray: Ground truth reconstruction volume with sparse binary features.
-    """
-    np.random.seed(42)
-    gt_recon = np.zeros(recon_shape, dtype=np.float32)
-    fill_rate = 0.05
-
-    y_pad = recon_shape[0] // 6
-    central_start = y_pad
-    central_end = recon_shape[0] - y_pad
-
-    row_size = recon_shape[1] * recon_shape[2]
-    num_ones_per_row = int(row_size * fill_rate)
-
-    for row_idx in range(central_start, central_end):
-        flat_row = gt_recon[row_idx].flatten()
-        positions_ones = np.random.choice(row_size, num_ones_per_row, replace=False)
-        flat_row[positions_ones] = 1.0
-        gt_recon[row_idx] = flat_row.reshape(recon_shape[1:])
-
-    return gt_recon
-
-
-
-# --- 3D Text Phantom Generator ---
-def gen_text_phantom(recon_shape, words, positions, font_path="DejaVuSans.ttf"):
-    """
-    Generate a 3D text phantom with binary word patterns embedded in specific slices.
-
-    Args:
-        recon_shape (tuple[int, int, int]): Shape of the phantom volume (num_rows, num_cols, num_slices).
-        words (list[str]): List of ASCII words to render.
-        positions (list[tuple[int, int, int]]): List of (row, col, slice) positions corresponding to each word.
-        array_size (int, optional): Size of the 2D square image used to render each word. Default is 256.
-        font_path (str, optional): Path to the TrueType font file. Default is "DejaVuSans.ttf".
-
-    Returns:
-        np.ndarray: A 3D numpy array of shape `recon_shape` containing the text phantom.
-    """
-    assert len(words) == len(positions), "Number of words must match number of positions."
-
-    array_size = np.minimum(recon_shape[1], recon_shape[2])
-
-    phantom = np.zeros(recon_shape, dtype=np.float32)
-    try:
-        font = ImageFont.truetype(font_path, size=20)
-    except OSError:
-        from pathlib import Path
-        fallback_paths = [
-            "/System/Library/Fonts/Supplemental/Arial.ttf",  # macOS fallback
-            "/Library/Fonts/Arial.ttf",  # Additional macOS path
-            "/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf",  # Linux fallback
-        ]
-        for fallback in fallback_paths:
-            if Path(fallback).exists():
-                font = ImageFont.truetype(fallback, size=20)
-                break
-        else:
-            raise FileNotFoundError(
-                f"Could not find a usable font. Tried the following paths:\n"
-                + "\n".join(fallback_paths)
-                + "\nPlease install one of these fonts or specify a valid font_path."
-            )
-
-    for word, (r, c, s) in zip(words, positions):
-        img = Image.new('L', (array_size, array_size), 0)
-        draw = ImageDraw.Draw(img)
-
-        text_box = draw.textbbox((0, 0), word, font=font)
-        text_width = text_box[2] - text_box[0]
-        text_height = text_box[3] - text_box[1]
-
-        x = (array_size - text_width) // 2
-        y = (array_size - text_height) // 2
-        draw.text((x, y), word, fill=1, font=font)
-
-        word_array = np.array(img.rotate(-90, expand=True).transpose(Image.FLIP_LEFT_RIGHT))
-        word_array = (word_array > 0).astype(np.float32)
-
-        # Crop or pad word_array to fit in the recon volume
-        r_start = max(0, r)
-        r_end = min(recon_shape[0], r + 1)
-        c_start = max(0, c - array_size // 2)
-        c_end = min(recon_shape[1], c_start + array_size)
-        s_start = max(0, s - array_size // 2)
-        s_end = min(recon_shape[2], s_start + array_size)
-
-        word_crop = word_array[:(c_end - c_start), :(s_end - s_start)]
-        phantom[r_start:r_end, c_start:c_end, s_start:s_end] = word_crop
-
-    return phantom
-
-
-# --- Translation Phantom Generator ---
-def gen_translation_phantom(option, recon_shape):
-    """
-    Generate a synthetic ground truth phantom based on the selected option.
-
-    Args:
-        option (str): Phantom type to generate. Options are 'dots' or 'text'.
-        recon_shape (tuple[int, int, int]): Shape of the reconstruction volume.
-
-    Returns:
-        np.ndarray: Generated phantom volume.
-    """
-    if option == 'dots':
-        return gen_dot_phantom(recon_shape)
-    elif option == 'text':
-        num_rows = recon_shape[0]
-        words = ["Purdue", "Presents", "Translation", "Tomography"]
-        positions = [
-            (num_rows // 5 * 1, recon_shape[1] // 2, recon_shape[2] // 2),
-            (num_rows // 5 * 2, recon_shape[1] // 2, recon_shape[2] // 2),
-            (num_rows // 5 * 3, recon_shape[1] // 2, recon_shape[2] // 2),
-            (num_rows // 5 * 4, recon_shape[1] // 2, recon_shape[2] // 2),
-        ]
-        return gen_text_phantom(recon_shape, words, positions)
-    else:
-        raise ValueError(f"Unsupported phantom option: {option}")
-
 
 
 def main():
@@ -254,7 +21,7 @@ def main():
     phantom_type = "text"   # Can be "dots" or "text"
 
     # Generate translation vectors
-    translation_vectors = gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing)
+    translation_vectors = mj.gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing)
 
     # Compute sinogram shape
     sino_shape = (translation_vectors.shape[0], num_det_rows, num_det_channels)
@@ -267,10 +34,10 @@ def main():
     # Print model parameters
     tct_model.print_params()
     # Display translation array
-    display_translation_vectors(translation_vectors, recon_shape)
+    mj.display_translation_vectors(translation_vectors, recon_shape)
 
     # Generate ground truth phantom
-    gt_recon = gen_translation_phantom(option='text', recon_shape=recon_shape)
+    gt_recon = mj.gen_translation_phantom(option=phantom_type, recon_shape=recon_shape)
 
     # View test sample
     mj.slice_viewer(gt_recon.transpose(0, 2, 1), title='Ground Truth Recon', slice_label='View', slice_axis=0)
@@ -292,8 +59,8 @@ def main():
                     title='Object (left) vs. MBIR reconstruction (right)', slice_axis=0)
 
     # Save as animated gifs
-    save_volume_as_gif(gt_recon, "gt_recon.gif", vmin=0, vmax=1)
-    save_volume_as_gif(recon, "mbir_recon.gif", vmin=0, vmax=1)
+    mj.save_volume_as_gif(gt_recon, "gt_recon.gif", vmin=0, vmax=1)
+    mj.save_volume_as_gif(recon, "mbir_recon.gif", vmin=0, vmax=1)
 
 
 if __name__ == '__main__':
