+++ mbirjax/vcd_utils.py
@@ -1,6 +1,7 @@
 import warnings
 from enum import Enum
 from typing import Union
+from PIL import Image, ImageDraw, ImageFont
 import numpy as np
 import jax.numpy as jnp
 import jax
@@ -13,17 +14,23 @@ class ObjectType(str, Enum):
     SHEPP_LOGAN = 'shepp-logan'
     CUBE = 'cube'
 
+
 class ModelType(str, Enum):
     PARALLEL = 'parallel'
     CONE = 'cone'
+    TRANSLATION = 'translation'
 
 
 def generate_demo_data(
     object_type: Union[ObjectType, str] = ObjectType.SHEPP_LOGAN,
     model_type: Union[ModelType, str] = ModelType.CONE,
     num_views: int = 64,
-    num_det_rows: int = 40,
-    num_det_channels: int = 128
+    num_det_rows: int = 96,
+    num_det_channels: int = 128,
+    num_x_translations: int = 7,
+    num_z_translations: int = 7,
+    x_spacing: float = 22,
+    z_spacing: float = 22
 ) -> (np.ndarray, np.ndarray):
     """
     Create a simple object and a sinogram for demonstration purposes.
@@ -38,10 +45,14 @@ def generate_demo_data(
 
     Args:
         object_type (str, optional): One of 'shepp-logan' or 'cube'.  Defaults to 'shepp-logan'.
-        model_type (str, optional): One of 'parallel' or 'cone'.  Defaults to 'cone'.
-        num_views (int, optional):  Number of views in the output sinogram.  Defaults to 64.
+        model_type (str, optional): One of 'parallel', 'cone', or 'translation'.  Defaults to 'cone'.
+        num_views (int, optional):  Number of views in the output sinogram.  Defaults to 64. Ignored when model_type is 'translation'
         num_det_rows (int, optional): Number of rows (vertical) in the output sinogram.  Defaults to 40.
         num_det_channels (int, optional): Number of channels (horizontal) in the output sinogram.  Defaults to 128.
+        num_x_translations (int, optional): Number of horizontal translations for translation mode.  Defaults to 7.
+        num_z_translations (int, optional): Number of vertical translations for translation mode.  Defaults to 7.
+        x_spacing (float, optional): Horizontal spacing between translations in ALU.  Defaults to 22.
+        z_spacing (float, optional): Vertical spacing between translations in ALU.  Defaults to 22.
 
     Returns:
         tuple: (object, sinogram, params)
@@ -53,25 +64,35 @@ def generate_demo_data(
     object_type = ObjectType(object_type)
     model_type = ModelType(model_type)
 
-    # For cone beam geometry, we need to describe the distances source to detector and source to rotation axis.
-    # np.Inf is an allowable value, in which case this is essentially parallel beam
-    source_detector_dist = 4 * num_det_channels
-    source_iso_dist = source_detector_dist
-
     start_angle = -np.pi
     end_angle = np.pi
 
     # Initialize model
-    sinogram_shape = (num_views, num_det_rows, num_det_channels)
-    angles = jnp.linspace(start_angle, end_angle, num_views, endpoint=False)
 
-    if model_type == ModelType.CONE:
+    if model_type == ModelType.PARALLEL:
+        sinogram_shape = (num_views, num_det_rows, num_det_channels)
+        angles = jnp.linspace(start_angle, end_angle, num_views, endpoint=False)
+        ct_model_for_generation = mj.ParallelBeamModel(sinogram_shape, angles)
+        params = {'angles': angles}
+    elif model_type == ModelType.CONE:
+        # For cone beam geometry, we need to describe the distances source to detector and source to rotation axis.
+        # np.Inf is an allowable value, in which case this is essentially parallel beam
+        source_detector_dist = 4 * num_det_channels
+        source_iso_dist = source_detector_dist
+        sinogram_shape = (num_views, num_det_rows, num_det_channels)
+        angles = jnp.linspace(start_angle, end_angle, num_views, endpoint=False)
         ct_model_for_generation = mj.ConeBeamModel(sinogram_shape, angles, source_detector_dist=source_detector_dist,
                                                    source_iso_dist=source_iso_dist)
         params = {'angles': angles, 'source_detector_dist': source_detector_dist, 'source_iso_dist': source_iso_dist}
-    elif model_type == ModelType.PARALLEL:
-        ct_model_for_generation = mj.ParallelBeamModel(sinogram_shape, angles)
-        params = {'angles': angles}
+    elif model_type == ModelType.TRANSLATION:
+        source_iso_dist = np.min(num_det_rows, num_det_channels) / 2
+        source_detector_dist = source_iso_dist
+        translation_vectors = gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing)
+        num_views = translation_vectors.shape[0]
+        sinogram_shape = (num_views, num_det_rows, num_det_channels)
+        ct_model_for_generation = mj.TranslationModel(sinogram_shape, translation_vectors, source_detector_dist=source_detector_dist,
+                                                      source_iso_dist=source_iso_dist)
+        params = {'translation_vectors': translation_vectors}
     else:
         raise ValueError(f'Invalid model type. Expected one of {[m.value for m in ModelType]}, got {model_type}')
 
@@ -92,6 +113,37 @@ def generate_demo_data(
     sinogram = np.asarray(sinogram)
 
     return phantom, sinogram, params
+
+
+def gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing):
+    """
+    Generate translation vectors for lateral (x) and axial (z) displacements.
+
+    Args:
+        num_x_translations (int): Number of x-direction translations
+        num_z_translations (int): Number of z-direction translations
+        x_spacing (float): Spacing between x translations in ALU
+        z_spacing (float): Spacing between z translations in ALU
+
+    Returns:
+        np.ndarray: Array of shape (num_views, 3) with translation vectors [dx, dy, dz]
+    """
+    num_views = num_x_translations * num_z_translations
+    translation_vectors = np.zeros((num_views, 3))
+
+    x_center = (num_x_translations - 1) / 2
+    z_center = (num_z_translations - 1) / 2
+
+    idx = 0
+    for row in range(num_z_translations):
+        for col in range(num_x_translations):
+            dx = (col - x_center) * x_spacing
+            dz = (row - z_center) * z_spacing
+            dy = 0
+            translation_vectors[idx] = [dx, dy, dz]
+            idx += 1
+
+    return translation_vectors
 
 
 def stitch_arrays(array_list, overlap_length, axis=2):
@@ -593,6 +645,131 @@ def generate_3d_shepp_logan_low_dynamic_range(phantom_shape, device=None):
     phantom = add_ellipsoid(phantom, grids, z_locations, 0, -0.1, 0, 0.046, 0.046, 0.046, intensity=0.1)
     phantom = add_ellipsoid(phantom, grids, z_locations, -0.08, -0.605, 0, 0.046, 0.023, 0.02, angle=0, intensity=0.1)
     phantom = add_ellipsoid(phantom, grids, z_locations, 0, -0.605, 0, 0.023, 0.023, 0.02, angle=0, intensity=0.1)
+
+    return phantom
+
+
+def gen_translation_phantom(option, recon_shape):
+    """
+    Generate a synthetic ground truth phantom based on the selected option.
+
+    Args:
+        option (str): Phantom type to generate. Options are 'dots' or 'text'.
+        recon_shape (tuple[int, int, int]): Shape of the reconstruction volume.
+
+    Returns:
+        np.ndarray: Generated phantom volume.
+    """
+    if option == 'dots':
+        return gen_dot_phantom(recon_shape)
+    elif option == 'text':
+        num_rows = recon_shape[0]
+        words = ["Purdue", "Presents", "Translation", "Tomography"]
+        positions = [
+            (num_rows // 5 * 1, recon_shape[1] // 2, recon_shape[2] // 2),
+            (num_rows // 5 * 2, recon_shape[1] // 2, recon_shape[2] // 2),
+            (num_rows // 5 * 3, recon_shape[1] // 2, recon_shape[2] // 2),
+            (num_rows // 5 * 4, recon_shape[1] // 2, recon_shape[2] // 2),
+        ]
+        return gen_text_phantom(recon_shape, words, positions)
+    else:
+        raise ValueError(f"Unsupported phantom option: {option}")
+
+
+def gen_dot_phantom(recon_shape):
+    """
+    Generate a synthetic ground truth reconstruction volume.
+
+    Args:
+        recon_shape (tuple[int, int, int]): Shape of the reconstruction volume.
+
+    Returns:
+        np.ndarray: Ground truth reconstruction volume with sparse binary features.
+    """
+    np.random.seed(42)
+    gt_recon = np.zeros(recon_shape, dtype=np.float32)
+    fill_rate = 0.05
+
+    y_pad = recon_shape[0] // 6
+    central_start = y_pad
+    central_end = recon_shape[0] - y_pad
+
+    row_size = recon_shape[1] * recon_shape[2]
+    num_ones_per_row = int(row_size * fill_rate)
+
+    for row_idx in range(central_start, central_end):
+        flat_row = gt_recon[row_idx].flatten()
+        positions_ones = np.random.choice(row_size, num_ones_per_row, replace=False)
+        flat_row[positions_ones] = 1.0
+        gt_recon[row_idx] = flat_row.reshape(recon_shape[1:])
+
+    return gt_recon
+
+
+def gen_text_phantom(recon_shape, words, positions, font_path="DejaVuSans.ttf"):
+    """
+    Generate a 3D text phantom with binary word patterns embedded in specific slices.
+
+    Args:
+        recon_shape (tuple[int, int, int]): Shape of the phantom volume (num_rows, num_cols, num_slices).
+        words (list[str]): List of ASCII words to render.
+        positions (list[tuple[int, int, int]]): List of (row, col, slice) positions corresponding to each word.
+        array_size (int, optional): Size of the 2D square image used to render each word. Default is 256.
+        font_path (str, optional): Path to the TrueType font file. Default is "DejaVuSans.ttf".
+
+    Returns:
+        np.ndarray: A 3D numpy array of shape `recon_shape` containing the text phantom.
+    """
+    assert len(words) == len(positions), "Number of words must match number of positions."
+
+    array_size = np.minimum(recon_shape[1], recon_shape[2])
+
+    phantom = np.zeros(recon_shape, dtype=np.float32)
+    try:
+        font = ImageFont.truetype(font_path, size=20)
+    except OSError:
+        from pathlib import Path
+        fallback_paths = [
+            "/System/Library/Fonts/Supplemental/Arial.ttf",  # macOS fallback
+            "/Library/Fonts/Arial.ttf",  # Additional macOS path
+            "/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf",  # Linux fallback
+        ]
+        for fallback in fallback_paths:
+            if Path(fallback).exists():
+                font = ImageFont.truetype(fallback, size=20)
+                break
+        else:
+            raise FileNotFoundError(
+                f"Could not find a usable font. Tried the following paths:\n"
+                + "\n".join(fallback_paths)
+                + "\nPlease install one of these fonts or specify a valid font_path."
+            )
+
+    for word, (r, c, s) in zip(words, positions):
+        img = Image.new('L', (array_size, array_size), 0)
+        draw = ImageDraw.Draw(img)
+
+        text_box = draw.textbbox((0, 0), word, font=font)
+        text_width = text_box[2] - text_box[0]
+        text_height = text_box[3] - text_box[1]
+
+        x = (array_size - text_width) // 2
+        y = (array_size - text_height) // 2
+        draw.text((x, y), word, fill=1, font=font)
+
+        word_array = np.array(img.rotate(-90, expand=True).transpose(Image.FLIP_LEFT_RIGHT))
+        word_array = (word_array > 0).astype(np.float32)
+
+        # Crop or pad word_array to fit in the recon volume
+        r_start = max(0, r)
+        r_end = min(recon_shape[0], r + 1)
+        c_start = max(0, c - array_size // 2)
+        c_end = min(recon_shape[1], c_start + array_size)
+        s_start = max(0, s - array_size // 2)
+        s_end = min(recon_shape[2], s_start + array_size)
+
+        word_crop = word_array[:(c_end - c_start), :(s_end - s_start)]
+        phantom[r_start:r_end, c_start:c_end, s_start:s_end] = word_crop
 
     return phantom
