+++ tests/test_vcd.py
@@ -4,7 +4,7 @@ import warnings
 import numpy as np
 import jax
 import jax.numpy as jnp
-import mbirjax
+import mbirjax as mj
 
 
 class TestVCD(unittest.TestCase):
@@ -16,10 +16,10 @@ class TestVCD(unittest.TestCase):
         """Set up before each test method."""
         np.random.seed(0)  # Set a seed to avoid variations due to partition creation.
         # Choose the geometry type
-        self.geometry_types = mbirjax._utils._geometry_types_for_tests
+        self.geometry_types = mj._utils._geometry_types_for_tests
         parallel_tolerances = {'nrmse': 0.15, 'max_diff': 0.38, 'pct_95': 0.04}
         cone_tolerances = {'nrmse': 0.19, 'max_diff': 0.56, 'pct_95': 0.05}
-        translation_tolerances = {'nrmse': 0.19, 'max_diff': 0.56, 'pct_95': 0.05}
+        translation_tolerances = {'nrmse': 0.8, 'max_diff': 1.05, 'pct_95': 0.03}
         self.all_tolerances = [parallel_tolerances, cone_tolerances, translation_tolerances]
         if len(self.geometry_types) != len(self.all_tolerances):
             raise IndexError('The list of geometry types does not match the list of test tolerances for the geometry types.')
@@ -38,12 +38,13 @@ class TestVCD(unittest.TestCase):
         # Initialize sinogram
         self.sinogram_shape = (self.num_views, self.num_det_rows, self.num_det_channels)
         self.angles = None
+        self.translation_vectors = None
 
     def tearDown(self):
         """Clean up after each test method."""
         pass
 
-    def set_angles(self, geometry_type):
+    def set_view_params(self, geometry_type):
         if geometry_type == 'cone':
             detector_cone_angle = 2 * np.arctan2(self.num_det_channels / 2, self.source_detector_dist)
         else:
@@ -52,35 +53,24 @@ class TestVCD(unittest.TestCase):
         end_angle = (np.pi + detector_cone_angle) * (1 / 2)
         self.angles = jnp.linspace(start_angle, end_angle, self.num_views, endpoint=False)
 
-        num_x_translations = 20
-        num_z_translations = 20
-        x_spacing = 1
-        z_spacing = 1
-        num_views = num_x_translations * num_z_translations
-        translation_vectors = np.zeros((num_views, 3))
+        num_x_translations = 7
+        num_z_translations = 7
+        x_spacing = 22
+        z_spacing = 22
 
-        x_center = (num_x_translations - 1) / 2
-        z_center = (num_z_translations - 1) / 2
-
-        idx = 0
-        for row in range(num_z_translations):
-            for col in range(num_x_translations):
-                dx = (col - x_center) * x_spacing
-                dz = (row - z_center) * z_spacing
-                dy = 0
-                translation_vectors[idx] = [dx, dy, dz]
-                idx += 1
+        # Generate translation vectors
+        translation_vectors = mj.gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing)
         self.translation_vectors = translation_vectors
 
     def get_model(self, geometry_type):
         if geometry_type == 'parallel':
-            ct_model = mbirjax.ParallelBeamModel(self.sinogram_shape, self.angles)
+            ct_model = mj.ParallelBeamModel(self.sinogram_shape, self.angles)
         elif geometry_type == 'cone':
-            ct_model = mbirjax.ConeBeamModel(self.sinogram_shape, self.angles,
+            ct_model = mj.ConeBeamModel(self.sinogram_shape, self.angles,
                                              source_detector_dist=self.source_detector_dist,
                                              source_iso_dist=self.source_iso_dist)
         elif geometry_type == 'translation':
-            ct_model = mbirjax.TranslationModel(self.sinogram_shape, self.translation_vectors,
+            ct_model = mj.TranslationModel(self.sinogram_shape, self.translation_vectors,
                                                 source_detector_dist=self.source_detector_dist,
                                                 source_iso_dist=self.source_iso_dist)
         else:
@@ -98,16 +88,17 @@ class TestVCD(unittest.TestCase):
         """
         Verify that the vcd reconstructions for a simple phantom are within tolerance
         """
-        self.set_angles(geometry_type)
+        self.set_view_params(geometry_type)
         ct_model = self.get_model(geometry_type)
 
         # Generate 3D Shepp Logan phantom
         print('  Creating phantom')
-        phantom = ct_model.gen_modified_3d_sl_phantom()
 
         if geometry_type == 'translation':
-            warnings.warn('test_vcd not implemented for translation mode.')
-            return
+            recon_shape = ct_model.get_params('recon_shape')
+            phantom = mj.gen_translation_phantom(option='text', recon_shape=recon_shape)
+        else:
+            phantom = ct_model.gen_modified_3d_sl_phantom()
 
         # Generate synthetic sinogram data
         print('  Creating sinogram')
@@ -138,7 +129,7 @@ class TestVCD(unittest.TestCase):
         with tempfile.NamedTemporaryFile('w') as file:
             filepath = file.name
             ct_model.save_recon_hdf5(filepath, recon, recon_dict)
-            loaded_recon, loaded_recon_dict, new_model = mbirjax.TomographyModel.load_recon_hdf5(str(filepath),
+            loaded_recon, loaded_recon_dict, new_model = mj.TomographyModel.load_recon_hdf5(str(filepath),
                                                                                                  recreate_model=True)
             loaded_notes = loaded_recon_dict['notes']
